fix(#120): axios 연결 오류 수정
