+++ src/pages/Student/MyClass/index.tsx
@@ -41,7 +41,6 @@ export default function MyClass() {
             <s.Card
               key={cls.id || `myclass-${idx}`}
               onClick={() => handleViewClass(cls.id)}
-              style={{ cursor: 'pointer' }}
             >
               <s.CardTitle>{cls.name}</s.CardTitle>
               <s.CardDescription>{cls.description || '설명이 없습니다.'}</s.CardDescription>
