+++ src/pages/Student/MyClass/styles.ts
@@ -59,7 +59,7 @@ export const Card = styled.div`
   flex-direction: column;
   justify-content: flex-start;  
   height: 10em;
-
+  cursor: pointer;
   &:hover {
     box-shadow: 0 4px 12px rgba(0,0,0,0.2);
   }
