+++ src/shared/types/Class/Assignment/Assignment.ts
@@ -1,3 +1,5 @@
+import {AssignmentAttachment } from '../Assignment/assignmentAttachment';
+
 // 과제 전체•단일 조회
 export interface AssignmentResponse {
   AssignmentAttachments: AssignmentAttachment[];
@@ -7,17 +9,6 @@ export interface AssignmentResponse {
   startDate: string; // 과제 시작일
   title: string; // 과제 제목
   userName: string; // 과제 생성자명
-  [property: string]: any;
-}
-
-// 과제 첨부
-export interface AssignmentAttachment {
-  contentType?: string;
-  originalFileName?: string; // 과제 파일이름
-  size?: number;
-  type: string; // 과제 타입
-  value: string; // 파일 경로 혹은 url
-  [property: string]: any;
 }
 
 // 과제 만들기
@@ -27,7 +18,6 @@ export interface AssignmentCreateRequest {
   end_date: string;
   start_date: string;
   title: string;
-  [property: string]: any;
 }
 
 // 과제 수정
@@ -36,17 +26,9 @@ export interface AssignmentUpdateRequest {
   end_date: string;
   start_date: string;
   title: string;
-  [property: string]: any;
 }
 
 // 과제 삭제
 export interface AssignmentDeleteResponse {
   message: string;
-  [property: string]: any;
-}
-
-export interface AssignmentFileType {
-  fileId: number | string;
-  fileName: string;
-  fileSize: number;
 }
